+++ Prelims/Problem 6/Solution.cpp
@@ -11,7 +11,7 @@ int main() {
     // Do a binary search to find the result
     int low = 0, high = n, res = 0;
     while (low <= high) {
-        int mid = (low + high) / 2;
+        ll mid = (low + high) / 2;
         if (mid * (mid + 1) / 2 <= n) {
             res = mid;
             low = mid + 1;
